feat: set customer subscription metadata flags on subscription creation and type change
